refactor: make builders types great again (#10026)
* refactor: make builders types great again

* fix: subcommands only type

---------
