+++ packages/builders/src/index.ts
@@ -49,9 +49,11 @@ export * from './interactions/slashCommands/options/user.js';
 export * from './interactions/slashCommands/mixins/ApplicationCommandNumericOptionMinMaxValueMixin.js';
 export * from './interactions/slashCommands/mixins/ApplicationCommandOptionBase.js';
 export * from './interactions/slashCommands/mixins/ApplicationCommandOptionChannelTypesMixin.js';
-export * from './interactions/slashCommands/mixins/ApplicationCommandOptionWithChoicesAndAutocompleteMixin.js';
+export * from './interactions/slashCommands/mixins/ApplicationCommandOptionWithAutocompleteMixin.js';
+export * from './interactions/slashCommands/mixins/ApplicationCommandOptionWithChoicesMixin.js';
 export * from './interactions/slashCommands/mixins/NameAndDescription.js';
 export * from './interactions/slashCommands/mixins/SharedSlashCommandOptions.js';
+export * from './interactions/slashCommands/mixins/SharedSubcommands.js';
 
 export * as ContextMenuCommandAssertions from './interactions/contextMenuCommands/Assertions.js';
 export * from './interactions/contextMenuCommands/ContextMenuCommandBuilder.js';
