+++ packages/builders/src/interactions/slashCommands/mixins/SharedSubcommands.ts
@@ -1,0 +1,171 @@
+import type {
+	LocalizationMap,
+	Permissions,
+	RESTPostAPIChatInputApplicationCommandsJSONBody,
+} from 'discord-api-types/v10';
+import {
+	assertReturnOfBuilder,
+	validateDMPermission,
+	validateDefaultMemberPermissions,
+	validateDefaultPermission,
+	validateLocalizationMap,
+	validateMaxOptionsLength,
+	validateNSFW,
+	validateRequiredParameters,
+} from '../Assertions.js';
+import type { ToAPIApplicationCommandOptions } from '../SlashCommandBuilder.js';
+import { SlashCommandSubcommandBuilder, SlashCommandSubcommandGroupBuilder } from '../SlashCommandSubcommands.js';
+
+/**
+ * This mixin holds symbols that can be shared in slash subcommands.
+ *
+ * @typeParam TypeAfterAddingSubcommands - The type this class should return after adding a subcommand or subcommand group.
+ */
+export class SharedSlashCommandSubcommands<
+	TypeAfterAddingSubcommands extends SharedSlashCommandSubcommands<TypeAfterAddingSubcommands>,
+> {
+	public readonly name: string = undefined!;
+
+	public readonly name_localizations?: LocalizationMap;
+
+	public readonly description: string = undefined!;
+
+	public readonly description_localizations?: LocalizationMap;
+
+	public readonly options: ToAPIApplicationCommandOptions[] = [];
+
+	/**
+	 * Sets whether the command is enabled by default when the application is added to a guild.
+	 *
+	 * @remarks
+	 * If set to `false`, you will have to later `PUT` the permissions for this command.
+	 * @param value - Whether or not to enable this command by default
+	 * @see {@link https://discord.com/developers/docs/interactions/application-commands#permissions}
+	 * @deprecated Use {@link SharedSlashCommandSubcommands.setDefaultMemberPermissions} or {@link SharedSlashCommandSubcommands.setDMPermission} instead.
+	 */
+	public setDefaultPermission(value: boolean) {
+		// Assert the value matches the conditions
+		validateDefaultPermission(value);
+
+		Reflect.set(this, 'default_permission', value);
+
+		return this;
+	}
+
+	/**
+	 * Sets the default permissions a member should have in order to run the command.
+	 *
+	 * @remarks
+	 * You can set this to `'0'` to disable the command by default.
+	 * @param permissions - The permissions bit field to set
+	 * @see {@link https://discord.com/developers/docs/interactions/application-commands#permissions}
+	 */
+	public setDefaultMemberPermissions(permissions: Permissions | bigint | number | null | undefined) {
+		// Assert the value and parse it
+		const permissionValue = validateDefaultMemberPermissions(permissions);
+
+		Reflect.set(this, 'default_member_permissions', permissionValue);
+
+		return this;
+	}
+
+	/**
+	 * Sets if the command is available in direct messages with the application.
+	 *
+	 * @remarks
+	 * By default, commands are visible. This method is only for global commands.
+	 * @param enabled - Whether the command should be enabled in direct messages
+	 * @see {@link https://discord.com/developers/docs/interactions/application-commands#permissions}
+	 */
+	public setDMPermission(enabled: boolean | null | undefined) {
+		// Assert the value matches the conditions
+		validateDMPermission(enabled);
+
+		Reflect.set(this, 'dm_permission', enabled);
+
+		return this;
+	}
+
+	/**
+	 * Sets whether this command is NSFW.
+	 *
+	 * @param nsfw - Whether this command is NSFW
+	 */
+	public setNSFW(nsfw = true) {
+		// Assert the value matches the conditions
+		validateNSFW(nsfw);
+		Reflect.set(this, 'nsfw', nsfw);
+		return this;
+	}
+
+	/**
+	 * Adds a new subcommand group to this command.
+	 *
+	 * @param input - A function that returns a subcommand group builder or an already built builder
+	 */
+	public addSubcommandGroup(
+		input:
+			| SlashCommandSubcommandGroupBuilder
+			| ((subcommandGroup: SlashCommandSubcommandGroupBuilder) => SlashCommandSubcommandGroupBuilder),
+	): TypeAfterAddingSubcommands {
+		const { options } = this;
+
+		// First, assert options conditions - we cannot have more than 25 options
+		validateMaxOptionsLength(options);
+
+		// Get the final result
+		const result = typeof input === 'function' ? input(new SlashCommandSubcommandGroupBuilder()) : input;
+
+		assertReturnOfBuilder(result, SlashCommandSubcommandGroupBuilder);
+
+		// Push it
+		options.push(result);
+
+		return this as unknown as TypeAfterAddingSubcommands;
+	}
+
+	/**
+	 * Adds a new subcommand to this command.
+	 *
+	 * @param input - A function that returns a subcommand builder or an already built builder
+	 */
+	public addSubcommand(
+		input:
+			| SlashCommandSubcommandBuilder
+			| ((subcommandGroup: SlashCommandSubcommandBuilder) => SlashCommandSubcommandBuilder),
+	): TypeAfterAddingSubcommands {
+		const { options } = this;
+
+		// First, assert options conditions - we cannot have more than 25 options
+		validateMaxOptionsLength(options);
+
+		// Get the final result
+		const result = typeof input === 'function' ? input(new SlashCommandSubcommandBuilder()) : input;
+
+		assertReturnOfBuilder(result, SlashCommandSubcommandBuilder);
+
+		// Push it
+		options.push(result);
+
+		return this as unknown as TypeAfterAddingSubcommands;
+	}
+
+	/**
+	 * Serializes this builder to API-compatible JSON data.
+	 *
+	 * @remarks
+	 * This method runs validations on the data before serializing it.
+	 * As such, it may throw an error if the data is invalid.
+	 */
+	public toJSON(): RESTPostAPIChatInputApplicationCommandsJSONBody {
+		validateRequiredParameters(this.name, this.description, this.options);
+
+		validateLocalizationMap(this.name_localizations);
+		validateLocalizationMap(this.description_localizations);
+
+		return {
+			...this,
+			options: this.options.map((option) => option.toJSON()),
+		};
+	}
+}
