+++ packages/builders/src/interactions/slashCommands/options/string.ts
@@ -2,7 +2,8 @@ import { s } from '@sapphire/shapeshift';
 import { ApplicationCommandOptionType, type APIApplicationCommandStringOption } from 'discord-api-types/v10';
 import { mix } from 'ts-mixer';
 import { ApplicationCommandOptionBase } from '../mixins/ApplicationCommandOptionBase.js';
-import { ApplicationCommandOptionWithChoicesAndAutocompleteMixin } from '../mixins/ApplicationCommandOptionWithChoicesAndAutocompleteMixin.js';
+import { ApplicationCommandOptionWithAutocompleteMixin } from '../mixins/ApplicationCommandOptionWithAutocompleteMixin.js';
+import { ApplicationCommandOptionWithChoicesMixin } from '../mixins/ApplicationCommandOptionWithChoicesMixin.js';
 
 const minLengthValidator = s.number.greaterThanOrEqual(0).lessThanOrEqual(6_000);
 const maxLengthValidator = s.number.greaterThanOrEqual(1).lessThanOrEqual(6_000);
@@ -10,7 +11,7 @@ const maxLengthValidator = s.number.greaterThanOrEqual(1).lessThanOrEqual(6_000)
 /**
  * A slash command string option.
  */
-@mix(ApplicationCommandOptionWithChoicesAndAutocompleteMixin)
+@mix(ApplicationCommandOptionWithAutocompleteMixin, ApplicationCommandOptionWithChoicesMixin)
 export class SlashCommandStringOption extends ApplicationCommandOptionBase {
 	/**
 	 * The type of this option.
@@ -67,4 +68,6 @@ export class SlashCommandStringOption extends ApplicationCommandOptionBase {
 	}
 }
 
-export interface SlashCommandStringOption extends ApplicationCommandOptionWithChoicesAndAutocompleteMixin<string> {}
+export interface SlashCommandStringOption
+	extends ApplicationCommandOptionWithChoicesMixin<string>,
+		ApplicationCommandOptionWithAutocompleteMixin {}
